pyproject.toml
[project]
name = "remote-claude"
version = "0.1.0"
description = "双端共享 Claude CLI 工具"
requires-python = ">=3.9"
dependencies = [
    "lark-oapi>=1.4.24",
    "python-dotenv>=1.0.0",
    "pyte>=0.8.0",
    "mixpanel>=5.0.0",
]

[project.scripts]
remote-claude = "remote_claude:main"
